+++ src/store/types/courses.ts
